+++ torch_geometric/loader/rearev_data_loader.py
@@ -1,12 +1,15 @@
-import torch
-from torch_geometric.data import Data, DataLoader
-from transformers import AutoTokenizer
 import json
-import os
+
 import numpy as np  # Added to fix undefined name 'np'
+import torch
+from transformers import AutoTokenizer
+
+from torch_geometric.data import Data, DataLoader
+
 
 class BasicDataLoader:
-    def __init__(self, config, word2id, relation2id, entity2id, tokenize, data_type="train"):
+    def __init__(self, config, word2id, relation2id, entity2id, tokenize,
+                 data_type="train"):
         self.batch_size = config['batch_size']
         print("Init called!")
         self.tokenize = tokenize
@@ -32,9 +35,9 @@ class BasicDataLoader:
         self.data = []
         file_path = f"{config['data_folder']}{data_type}.json"
         print(f"Loading data from {file_path}...")
-        with open(file_path, 'r') as f:
+        with open(file_path) as f:
             lines = len(f.readlines())
-        with open(file_path, 'r') as f:
+        with open(file_path) as f:
             print("Number of lines: ", lines)
             iter = 0
             for line in f:
@@ -48,10 +51,14 @@ class BasicDataLoader:
 
     def _prepare_data(self):
         self.graphs = [self._create_graph(sample) for sample in self.data]
-        self.data_loader = DataLoader(self.graphs, batch_size=self.batch_size, shuffle=True)
+        self.data_loader = DataLoader(self.graphs, batch_size=self.batch_size,
+                                      shuffle=True)
 
     def _create_graph(self, sample):
-        entity_map = {ent: idx for idx, ent in enumerate(sample.get('entities', []))}
+        entity_map = {
+            ent: idx
+            for idx, ent in enumerate(sample.get('entities', []))
+        }
         x = torch.zeros(len(entity_map), self.num_relations)
         edges, edge_attrs = [], []
 
@@ -69,7 +76,8 @@ class BasicDataLoader:
         return Data(x=x, edge_index=edge_index, edge_attr=edge_attr)
 
     def reset_batches(self):
-        self.data_loader = DataLoader(self.graphs, batch_size=self.batch_size, shuffle=True)
+        self.data_loader = DataLoader(self.graphs, batch_size=self.batch_size,
+                                      shuffle=True)
 
     def build_rel_words(self, tokenize):
         # Tokenizes relation surface forms.
@@ -93,15 +101,16 @@ class BasicDataLoader:
                 except Exception:  # Changed bare except to except Exception
                     words = ['UNK']
                     rel_words.append(words)
-                    pass
                 # words = fields[-2].split('_') + fields[-1].split('_')
 
         self.max_rel_words = max_rel_words
         if tokenize == 'lstm':
-            self.rel_texts = np.full((self.num_kb_relation + 1, self.max_rel_words),
-                                     len(self.word2id), dtype=int)
-            self.rel_texts_inv = np.full((self.num_kb_relation + 1, self.max_rel_words),
-                                         len(self.word2id), dtype=int)
+            self.rel_texts = np.full(
+                (self.num_kb_relation + 1, self.max_rel_words),
+                len(self.word2id), dtype=int)
+            self.rel_texts_inv = np.full(
+                (self.num_kb_relation + 1, self.max_rel_words),
+                len(self.word2id), dtype=int)
             for rel_id, tokens in enumerate(rel_words):
                 for j, word in enumerate(tokens):
                     if j < self.max_rel_words:
@@ -129,26 +138,23 @@ class BasicDataLoader:
 
             tokenizer = AutoTokenizer.from_pretrained(tokenizer_name)
             pad_val = tokenizer.convert_tokens_to_ids(tokenizer.pad_token)
-            self.rel_texts = np.full((self.num_kb_relation + 1, self.max_rel_words),
-                                     pad_val, dtype=int)
-            self.rel_texts_inv = np.full((self.num_kb_relation + 1, self.max_rel_words),
-                                         pad_val, dtype=int)
+            self.rel_texts = np.full(
+                (self.num_kb_relation + 1, self.max_rel_words), pad_val,
+                dtype=int)
+            self.rel_texts_inv = np.full(
+                (self.num_kb_relation + 1, self.max_rel_words), pad_val,
+                dtype=int)
 
             for rel_id, words in enumerate(rel_words):
-                tokens = tokenizer.encode_plus(
-                    text=' '.join(words),
-                    max_length=self.max_rel_words,
-                    pad_to_max_length=True,
-                    return_attention_mask=False,
-                    truncation=True
-                )
+                tokens = tokenizer.encode_plus(text=' '.join(words),
+                                               max_length=self.max_rel_words,
+                                               pad_to_max_length=True,
+                                               return_attention_mask=False,
+                                               truncation=True)
                 tokens_inv = tokenizer.encode_plus(
-                    text=' '.join(words[::-1]),
-                    max_length=self.max_rel_words,
-                    pad_to_max_length=True,
-                    return_attention_mask=False,
-                    truncation=True
-                )
+                    text=' '.join(words[::-1]), max_length=self.max_rel_words,
+                    pad_to_max_length=True, return_attention_mask=False,
+                    truncation=True)
                 self.rel_texts[rel_id] = np.array(tokens['input_ids'])
                 self.rel_texts_inv[rel_id] = np.array(tokens_inv['input_ids'])
 
@@ -172,13 +178,14 @@ def load_data(config, tokenize):
     print("Dictionaries loaded!")
 
     loaders = {
-        data_type: SingleDataLoader(config, word2id, relation2id, entity2id, tokenize, data_type)
+        data_type:
+        SingleDataLoader(config, word2id, relation2id, entity2id, tokenize,
+                         data_type)
         for data_type in ['train', 'dev', 'test']
     }
 
     return {
-        **loaders,
-        "entity2id": entity2id,
+        **loaders, "entity2id": entity2id,
         "relation2id": relation2id,
         "word2id": word2id,
         "num_word": AutoTokenizer.from_pretrained(tokenize)